+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [0.8.7] - 2022-05-04
+
+* v5: Account for property type byte in property length when encoding Subscribe packet
+
 ## [0.8.6] - 2022-xx-xx
 
 * v5: Add Router::finish() helper method, it converts router to service factory